Update vmss-mz.ps1
VMSS Flex is what the PG will be pushing moving forward especially to replace AvSet. One thing I'm unsure is how to enable max spreading in PS or is that set by default? There is param for -PlatformFaultDomainCount, and I wonder if setting it as 0 means max spreading across fault domains.